% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estMCS.reg.R
\name{estMCS.reg}
\alias{estMCS.reg}
\title{Estimation of model confidence sets for linear regressions using an in-sample
loss measure.}
\usage{
estMCS.reg(data, models, B = 1000, l = 2)
}
\arguments{
\item{data}{A matrix containing the data set to be used.}

\item{models}{A list with one entry for each model specifying in a numerical
vector the columns of matrix \code{data} to be used in that model.}

\item{B}{the number of bootstrap samples.}

\item{l}{the block length used in the moving-block bootstrap.}
}
\value{
...
}
\description{
The function allows to estimate a model confidence set a la Hansen, Lunde
and Nason (2011) for the case of linear regression models. A matrix is
returned that lists the entered models with their likelihood, estimated
efficient degrees of freedom, KLIC, BIC, AIC and the MCS p-values associated
with each information criterion.
}
\details{
...
}
\references{
Hansen, P. R., Lunde, A., Nason, J. M. 2011. "The Model
  Confidence Set", \emph{Econometrica}, \bold{79(2)}, 453 - 497
}
\seealso{
\code{\link{estMCS}}, \code{\link{estMCS.quick}}
}
\author{
Niels Aka
}
\keyword{htest}
\keyword{models}
\keyword{robust}
