% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconf-package.R
\docType{package}
\name{modelconf-package}
\alias{modelconf}
\alias{modelconf-package}
\title{Model Confidence Sets}
\description{
An alternative implementation of the Model Confidece Set (MCS).
}
\details{
Three functions implement the algorithms in Hansen, Lunde, Nason (2011).
\itemize{
\item \code{\link{estMCS}}
\item \code{\link{estMCS.quick}}
\item \code{\link{estMCS.reg}}
}
}
\references{
Hansen, P. R., Lunde, A., Nason, J. M. 2011. "The Model
Confidence Set", \emph{Econometrica}, \bold{79(2)}, 453 - 497
}
\author{
\strong{Maintainer}: Niels Aka \email{nielsaka@outlook.com}

Authors:
\itemize{
  \item Rolf Tschernig
}

}
