% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconf-package.R
\docType{data}
\name{SW_infl4cast}
\alias{SW_infl4cast}
\title{Forecasts based on models found in Stock and Watson (1999).}
\format{A data frame with 168 observations on the following 21 variables.
\describe{ \item{list("Obs")}{a numeric vector} \item{list("NoChange.A")}{a
numeric vector} \item{list("NoChange.B")}{a numeric vector}
\item{list("uniar")}{a numeric vector} \item{list("dtip")}{a numeric vector}
\item{list("dtgmpyq")}{a numeric vector} \item{list("dtmsmtq")}{a numeric
vector} \item{list("dtlpnag")}{a numeric vector} \item{list("ipxmca")}{a
numeric vector} \item{list("hsbp")}{a numeric vector}
\item{list("lhmu25")}{a numeric vector} \item{list("ip")}{a numeric vector}
\item{list("gmpyq")}{a numeric vector} \item{list("msmtq")}{a numeric
vector} \item{list("lpnag")}{a numeric vector} \item{list("dipxmca")}{a
numeric vector} \item{list("dhsbp")}{a numeric vector}
\item{list("dlhmu25")}{a numeric vector} \item{list("dlhur")}{a numeric
vector} \item{list("lhur")}{a numeric vector} }}
\source{
Hansen, P. R., Lunde, A., Nason, J. M. 2011. "The Model Confidence
Set", \emph{Econometrica}, \bold{79(2)}, 453 - 497
}
\description{
This data set is taken from an article by Hansen, Lunde and Nason (2011). It
allows to apply the functions of this package and to compare the results
with the original outcomes in the paper.
}
\details{
...
}
\examples{

data(SW_infl4cast)

}
\references{
Stock and Watson (....)
}
\keyword{datasets}
