% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelconf-package.R
\docType{data}
\name{TR_regs}
\alias{TR_regs}
\title{Observational data for Taylor Rule regressions.}
\format{A data frame with 148 observations on the following 21 variables.
\describe{ \item{list("X")}{a factor} \item{list("Constant")}{a numeric
vector} \item{list("Trend")}{a numeric vector} \item{list("FFR")}{a numeric
vector} \item{list("dFFR")}{a numeric vector} \item{list("GDP")}{a numeric
vector} \item{list("lnGDP")}{a numeric vector} \item{list("GDPD.I")}{a
numeric vector} \item{list("GDPC.I")}{a numeric vector}
\item{list("HPCE.I")}{a numeric vector} \item{list("CPCE.I")}{a numeric
vector} \item{list("HCPI.I")}{a numeric vector} \item{list("CCPI.I")}{a
numeric vector} \item{list("GAP_HP")}{a numeric vector}
\item{list("GAP_Gr")}{a numeric vector} \item{list("GAP_Tr")}{a numeric
vector} \item{list("GAP_BK")}{a numeric vector} \item{list("Unemp")}{a
numeric vector} \item{list("UMP.gap")}{a numeric vector}
\item{list("Lab.Cost")}{a numeric vector} \item{list("GDPD.I.GAP_HP")}{a
numeric vector} }}
\source{
Hansen, Lunde, Nason (2011) url ...
}
\description{
This data set is taken from an article by Hansen, Lunde and Nason (2011). It
allows to apply the functions of this package and to compare the results
with the original outcomes in the paper.
}
\examples{

data(TR_regs)

}
\references{
Hansen, Lunde, Nason (2011)
}
\keyword{datasets}
